[1] Adjustments of CMake port to Boost libraries. [2] Update of build script to better handle testing for installed CMake [3] Continued cleaning up of imaging routines.
git-svn-id: http://usg.lofar.org/svn/code/trunk@1947 e3a606d5-b720-0410-b6ad-cb58f335abf8

+++ src/CR-Tools/implement/IO/test/tFalckeGorhamGalacticNoise.cc
@@ -76,7 +76,7 @@ int main ()
   for (long i=0; i<numpoints; ++i) { frequencies(i)=i*freqstep; }
   Test.getNoise(frequencies,noise);
   
-  for (long i=0; i<noise.nelements(); ++i)
+  for (long i=0; i<long(noise.nelements()); ++i)
   {
     cout << frequencies(i) << "\t" << noise(i) << "\n";
   }
@@ -86,7 +86,7 @@ int main ()
   cout << endl;
 
   double timeintegral = 0.0;
-  for (long i=0; i<timeseries.nelements(); ++i)
+  for (long i=0; i<long(timeseries.nelements()); ++i)
   {
     cout << timeseries(i) << "\n";
     timeintegral+=timeseries(i)*timeseries(i)/impedance*timestep;
